requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.6
django-cors-middleware==1.3.1
django-filter==2.0.0
django-oauth-toolkit==1.2.0
djangorestframework==3.8.2
idna==2.7
Markdown==2.6.11
oauthlib==2.1.0
PyGithub==1.40
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
